feat(ws): add stmt API
Now only support:

```C
WS_STMT *ws_stmt_init(const WS_TAOS *taos);

int ws_stmt_prepare(WS_STMT *stmt, const char *sql, unsigned long length);

int ws_stmt_set_tbname(WS_STMT *stmt, const char *name);

int ws_stmt_set_tags(WS_STMT *stmt, const WS_MULTI_BIND *bind, uint32_t len);

int ws_stmt_bind_param_batch(WS_STMT *stmt, const WS_MULTI_BIND *bind, uint32_t len);

int ws_stmt_add_batch(WS_STMT *stmt);

int ws_stmt_execute(WS_STMT *stmt, int32_t *affected_rows);

int ws_stmt_affected_rows(WS_STMT *stmt);

void ws_stmt_close(WS_STMT *stmt);

```

Ref: [TD-17318](https://jira.taosdata.com:18080/browse/TD-17318)
