+++ taos-ws-sys/Cargo.toml
@@ -11,6 +11,7 @@ crate-type = ["cdylib", "staticlib"]
 
 [dependencies]
 anyhow = "1"
+serde_json = "1"
 taos-error = { path = "../taos-error" }
 taos-query = { path = "../taos-query" }
 taos-ws = { path = "../taos-ws", default-features = false }
